% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{design}
\alias{design}
\title{Design}
\usage{
design(x, w, tag = list())
}
\arguments{
\item{x}{a row matrix of points.}

\item{w}{a vector of weights.
Length shall be equal to the number of rows in \code{x} and sum shall be equal to \code{1}.}

\item{tag}{a list containing additional information about the design.}
}
\value{
\code{design} returns an object of \code{class} \code{"desigh"}.
An object of class \code{"desigh"} is a list containing at least this function's arguments.
}
\description{
\code{design} creates a custom design object.
}
\examples{
## see examples for param

}
\seealso{
\code{\link{Wynn}}, \code{\link{reduce}}, \code{\link{getM}}, \code{\link{plot.desigh}}, \code{\link{Defficiency}}, \code{\link{update.param}}
}

