% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docopulae.R
\docType{package}
\name{docopulae}
\alias{docopulae}
\alias{docopulae-package}
\title{Design of Experiments with Copulas}
\description{
docopulae offers a set of functions to define parametric models by their
joint density (likelihood), evaluate the expected Fisher information,
find D- and Ds-optimal designs and eventually compute common statistics.
A crucial aspect is modularity, so the user may replace any function
by a specialized solution. The package also consists of a number of
convenience functions for general purpose and specific to the task.
}
\details{
TODO
}

