% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{Dsensitivity}
\alias{Dsensitivity}
\title{D Sensitivity}
\usage{
Dsensitivity(sNames = NULL, names = NULL, defaults = list(x = NULL, desw =
  NULL, desx = NULL, mod = NULL))
}
\description{
D Sensitivity
}

