% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{expr2f}
\alias{expr2f}
\title{Expression To Function}
\usage{
expr2f(x, map = NULL, yMap = NULL, thetaMap = NULL)
}
\arguments{
\item{x}{an expression.}

\item{map}{a named list of names defining left assignments (\code{a="b"} := \code{a <- b}).}

\item{yMap}{like \code{map} but items \code{a=1} resolve to \code{a <- y[1]}.}

\item{thetaMap}{like \code{yMap} for \code{theta} with single element access \code{[[}.}
}
\value{
\code{expr2f} returns \code{function(y, theta, ...)}, where \code{theta} is a list of parameters.
It evaluates expression \code{x}.
}
\description{
Turns an expression into \code{function(y, theta, ...)}.
}
\examples{
library(docopulae)
library(copula)

## build f
margins = list(alist(pdf=dnorm(y1, mu1, sd1),
                     cdf=pnorm(y1, mu1, sd1)),
               alist(pdf=dnorm(y2, mu2, sd2),
                     cdf=pnorm(y2, mu2, sd2)))
C = claytonCopula()
f = buildf(margins, C)
f

## turn f into function
expr2f(f, yMap=list(y1=1, y2=2), thetaMap=list(alpha='alpha', mu1='mu1', mu2='mu2', sd1='sd1', sd2='sd2'))
}
\seealso{
\code{\link{buildf}}, \code{\link{fisherI}}
}

