% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{FedorovWynn}
\alias{FedorovWynn}
\title{Fedorov Wynn Algorithm}
\usage{
FedorovWynn(mod, names = NULL, tolAbs = Inf, tolRel = 1e-04,
  maxIter = 10000)
}
\arguments{
\item{mod}{an object of \code{class} \code{"param"}.}

\item{names}{a vector of names or indices, the subset of parameters to optimize for.}

\item{tolAbs}{the absolute tolerance for the upper bound of the sensitivity.}

\item{tolRel}{the relative tolerance with respect to the number of parameters.}

\item{maxIter}{the maximum number of iterations.}
}
\value{
\code{FedorovWynn} returns an object of \code{class} \code{"design"}.
An object of class \code{"design"} is a list containing the following components:
\itemize{
\item mod: argument
\item x: row matrix of design points, in this case \code{mod$x}.
\item w: a numeric vector of weights for each design point.
\item sens: a numeric vector of sensitivity for each design point.
\item args: a list of arguments.
\item adds: a list of additional (runtime) information.
}
}
\description{
\code{FedorovWynn} finds a D- or Ds-optimal design using the Fedorov-Wynn algorithm.
}
\details{
TODO shortly describe what is done and why (reference to paper).
}
\examples{
#TODO
}
\seealso{
\code{\link{param}}, \code{\link{reduce}}
}

