% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{grow.grid}
\alias{grow.grid}
\title{Grow Grid}
\usage{
grow.grid(x)
}
\arguments{
\item{x}{a list of vectors.}
}
\value{
\code{grow.grid} returns a data frame like \code{expand.grid}.
}
\description{
\code{grow.grid} creates a data frame like \code{expand.grid}.
The order of rows is adjusted to represent a growing grid with respect to resolution.
}
\seealso{
\code{\link{update.param}}
}
