% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{integrateA}
\alias{integrateA}
\title{Integrate Alternative}
\usage{
integrateA(f, lower, upper, ..., subdivisions = 100L,
  rel.tol = .Machine$double.eps^0.25, abs.tol = rel.tol,
  stop.on.error = TRUE, keep.xy = FALSE, aux = NULL)
}
\description{
A tolerance wrapper for \code{stats::integrate}. It allows \code{integrate} to reach the maximum number of subdivisions.
}
\details{
See \code{stats::integrate}.
}
\seealso{
\code{\link[stats]{integrate}} in \pkg{stats}
}

