% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_expandSpace}
\alias{nint_expandSpace}
\title{Expand Space}
\usage{
nint_expandSpace(x)
}
\arguments{
\item{x}{some space or a list structure of spaces.}
}
\value{
\code{nint_expandSpace} returns a list of spaces.
Each space is a true subspace.
}
\description{
\code{nint_expandSpace} expands some space or list structure of spaces to a list of true subspaces.
}
\examples{
s = nint_space(list(nint_intvDim(1, 2),
                    nint_intvDim(3, 4)),
               list(nint_intvDim(-Inf, 0),
                    nint_gridDim(c(0)),
                    nint_intvDim(0, Inf))
               )
s
nint_expandSpace(s)
}
\seealso{
\code{\link{nint_space}}
}

