% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_funcDim}
\alias{nint_funcDim}
\title{Function Dimension}
\usage{
nint_funcDim(x)
}
\arguments{
\item{x}{\code{function(x)}, where \code{x} is the partially realized point in the space.
It shall return an object of type \code{nint_intvDim} or a vector.}
}
\value{
\code{nint_scatDim} returns its argument with the dimension type attribute set to \code{nint_TYPE_FUNC_DIM}.
}
\description{
\code{nint_funcDim} defines a functionally dependent dimension.
It shall depend solely on the previous dimensions.
}
\details{
Obviously if \code{x} returns an object of type \code{nint_intvDim} the dimension is continuous, and discrete otherwise.

As the argument to \code{x} is only partially defined the user has to ensure that the function solely depends on values up to the current dimension.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}

