% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_gridDim}
\alias{nint_gridDim}
\title{Grid Dimension}
\usage{
nint_gridDim(x)
}
\arguments{
\item{x}{a vector of any type.}
}
\value{
\code{nint_scatDim} returns its argument with the dimension type attribute set to \code{nint_TYPE_GRID_DIM}.
}
\description{
\code{nint_gridDim} is defined by a vector of values which in combination with other grid dimensions creates a dense grid.
}
\details{
Imagine using expand.grid to create a matrix where each row defines a point.
This is the role \code{nint_gridDim} plays.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}

