% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_integrateNFunc}
\alias{nint_integrateNFunc}
\alias{nint_integrateNFunc_recursive}
\title{Integrate N Function}
\usage{
nint_integrateNFunc(f, funcs, x0, i0, ...)

nint_integrateNFunc_recursive(integrateIntv)
}
\arguments{
\item{integrateIntv}{\code{function(f, lowerLimit, upperLimit, ...)}, which performs one dimensional integration.}

\item{f}{the scalar-valued function (integrand) to be integrated.}

\item{funcs}{the list of functions. See \code{\link{nint_funcDim}}.}

\item{x0}{the partially realized point.}

\item{i0}{the vector of indices where the function dimensions fit into \code{x0}.}

\item{...}{other arguments passed to \code{f}.}
}
\value{
\code{nint_integrateNFunc} returns a single numeric.
}
\description{
Inferface to the integration over function dimensions.
}
\details{
\code{nint_integrateNCube} is a reference to the function that \code{nint_integrate} calls with exactly these arguments to integrate over function dimensions.
See examples below on how to replace it with a different function.

\code{nint_integrateNFunc_recursive} returns a recursive implementation which directly sums over discrete dimensions and uses \code{integrateIntv} otherwise. In conjunction with \code{stats::integrate} this is the default.
}
\examples{
library(docopulae)

## volume of sphere
s = nint_space(
    nint_intvDim(-1, 1),
    nint_funcDim(function(x) nint_intvDim(c(-1, 1) * sin(acos(x[1])) )),
    nint_funcDim(function(x) {
        r = sin(acos(x[1]))
        nint_intvDim(c(-1, 1) * r*cos(asin(x[2] / r)))
    }))
nint_integrate(function(x) 1, s) # 4*pi/3

## replace nint_integrateNFunc
f = function(f, funcs, x0, i0, ...) {
    print(list(f=f, funcs=funcs, x0=x0, i0=i0, ...))
    # disregard unknown structure of f
    stop('do something')
}
unlockBinding('nint_integrateNFunc', environment(nint_integrate))
assign('nint_integrateNFunc', f, envir=environment(nint_integrate))

## integrate with replacement
try(nint_integrate(function(x) 1, s))
}
\seealso{
\code{\link{nint_integrate}}

\code{\link[stats]{integrate}} in \pkg{stats}
}

