% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_intvDim}
\alias{nint_intvDim}
\title{Interval Dimension}
\usage{
nint_intvDim(x, b = NULL)
}
\arguments{
\item{x}{either a vector containing lower and upper limit, or a single value specifying only the lower limit.}

\item{b}{the upper limit if \code{x} is the lower limit.}
}
\value{
\code{nint_intvDim} returns a vector of length 2 with the dimension type attribute set to \code{nint_TYPE_INTV_DIM}.
}
\description{
\code{nint_intvDim} defines a continous range with a fixed lower and upper limit.
The limits may be (negative) \code{Inf}.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}, \code{\link{nint_transform}}
}

