% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_ispace}
\alias{nint_ispace}
\title{Create Integration Space}
\usage{
nint_ispace(x)
}
\arguments{
\item{x}{a true subspace.}
}
\value{
\code{nint_ispace} returns a list.
}
\description{
\code{nint_ispace} transforms a true subspace into a data structure which is used by \code{nint_integrate} to efficiently integrate over the entire space.
}
\details{
Assumes interchangeability of dimensions (except function dimensions).
}
\examples{
library(docopulae)

s = nint_space(nint_scatDim(c(1, 3, 5)),
               nint_gridDim(1:3),
               nint_scatDim(c(2, 4, 6)),
               nint_gridDim(4:5),
               nint_funcDim(function(x) x[1:4]),
               nint_intvDim(-Inf, 3),
               nint_gridDim(6:9),
               nint_intvDim(-Inf, Inf),
               nint_intvDim(-pi, Inf))
s

nint_ispace(s)
# i: indices
# g: data
}
\seealso{
\code{\link{nint_space}}, \code{\link{nint_integrate}}
}

