% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_scatDim}
\alias{nint_scatDim}
\title{Scattered Dimension}
\usage{
nint_scatDim(x)
}
\arguments{
\item{x}{a vector of any type.}
}
\value{
\code{nint_scatDim} returns its argument with the dimension type attribute set to \code{nint_TYPE_SCAT_DIM}.
}
\description{
\code{nint_scatDim} is defined by a vector of values which in combination with other scatter dimensions creates a sparse grid.
}
\details{
Imagine using cbind to create a matrix where each row defines a point.
This is the role \code{nint_scatDim} plays.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}

