% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_space}
\alias{nint_space}
\title{Space}
\usage{
nint_space(...)
}
\arguments{
\item{...}{dimensions each of which may be an actual dimension object or a list structure of dimension objects.}
}
\value{
\code{nint_space} returns an object of \code{class} \code{"nint_space"}.
}
\description{
\code{nint_space} defines an n-dimensional space as a list of dimensions.
A space may contain subspaces.
A space without subspaces is called true subspace.
}
\details{
If a space contains at least one list structure of dimension objects it consists of subspaces each defined by a combination of dimension objects along all dimensions.
See \code{\link{nint_expandSpace}} on how to extract them.
}
\examples{
library(docopulae)

s = nint_space(nint_gridDim(1:4),
               list(nint_intvDim(-Inf, -3), nint_intvDim(3, Inf)),
               list(list(list(nint_scatDim(c(1, 2))), nint_scatDim(c(2, 3))),
                    nint_scatDim(c(3, 4)) ))
s
}
\seealso{
\code{\link{nint_scatDim}}, \code{\link{nint_gridDim}}, \code{\link{nint_intvDim}}, \code{\link{nint_funcDim}}, \code{\link{nint_transform}}, \code{\link{nint_integrate}}, \code{\link{nint_expandSpace}}
}

