% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nint.R
\docType{data}
\name{nint_TYPE}
\alias{nint_TYPE}
\alias{nint_TYPE_FUNC_DIM}
\alias{nint_TYPE_GRID_DIM}
\alias{nint_TYPE_INTV_DIM}
\alias{nint_TYPE_SCAT_DIM}
\title{Dimension Type Attribute Values}
\format{integer}
\usage{
nint_TYPE_SCAT_DIM # = 1

nint_TYPE_GRID_DIM # = 2

nint_TYPE_INTV_DIM # = 3

nint_TYPE_FUNC_DIM # = 4
}
\description{
A dimension object is identified by its dimension type attribute \code{"nint_dtype"}.
On creation it is set to one of the following.
See dimension types in "See Also" below.
}
\seealso{
\code{\link{nint_scatDim}}, \code{\link{nint_gridDim}}, \code{\link{nint_intvDim}}, \code{\link{nint_funcDim}}, \code{\link{nint_space}}
}
\keyword{datasets}

