% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{param}
\alias{param}
\title{Parametric Model}
\usage{
param(fisherIf, dDim)
}
\arguments{
\item{fisherIf}{\code{function(x, ...)}, where \code{x} is a numeric vector, usually a point from the design space.
It shall evaluate to the Fisher information.}

\item{dDim}{length of \code{x}, usually the dimensionality of the design space.}
}
\value{
\code{param} returns an object of \code{class} \code{"param"}.
An object of class \code{"param"} is a list containing the following components:
\itemize{
\item fisherIf: argument
\item dDim: argument
\item x: row matrix of points where \code{fisherIf} has already been evaluated.
\item fisherI: list of Fisher information matrices for each row in \code{x} respectively.
}
}
\description{
\code{param} creates an initial parametric model object.
Unlike other model statements this function does not perform any computation.
}
\seealso{
\code{\link{fisherI}}, \code{\link{update.param}},  \code{\link{FedorovWynn}}
}

