% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{plot_design}
\alias{plot_design}
\title{Plot Design}
\usage{
plot_design(des, ..., margins = NULL, wDes = NULL, plus = T,
  circles = F, border = c(0.1, 0.1, 0, 0.1), sensArgs = list(),
  wArgs = list())
}
\arguments{
\item{des}{some design.}

\item{...}{other arguments passed to plot.}

\item{margins}{a vector of indices.
Defaults to \code{1}.}

\item{wDes}{a design from which to take the weights.
Defaults to argument \code{des}.
See \code{reduce}.}

\item{plus}{add plus symbols to the sensitivity curve.}

\item{circles}{draw weights as circles instead of as bars.}

\item{border}{\code{c(bottom, left, top, right)}, relative margins to add if drawing circles.}

\item{sensArgs}{a list of arguments for drawing the sensitivity.}

\item{wArgs}{a list of arguments for drawing the weights.}
}
\description{
\code{plot_design} visualizes the weights and sensitivities of some design.
Designs with more than one dimension are projected to a specified margin.
}
\details{
The diameter of each circle is proportional to its weight.
}
\examples{
#TODO
}
\seealso{
\code{\link{FedorovWynn}}, \code{\link{reduce}}
}

