% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nint.R
\name{print.nint_space}
\alias{print.nint_space}
\title{Print Space}
\usage{
\method{print}{nint_space}(x, ...)
}
\arguments{
\item{x}{a space.}

\item{...}{ignored.}
}
\description{
\code{print.nint_space} prints a space in a convenient way.
}
\details{
Each line represents a dimension.
Format: "<dim idx>: <dim repr>".
Each dimension has its own representation which should be easy to understand.
\code{nint_scatDim} representations are marked by \code{"s()"}.
}
\seealso{
\code{\link{nint_space}}
}
