% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{reduce}
\alias{reduce}
\title{Reduce Design}
\usage{
reduce(des, distMax, wMin = 1e-06)
}
\arguments{
\item{des}{an object of \code{class} \code{"design"}.}

\item{distMax}{maximum euclidean distance between points to be merged.}

\item{wMin}{minimum weight a significant design point shall have.}
}
\value{
\code{reduce} returns an object of \code{class} \code{"design"}.
See \code{FedorovWynn} for its structural definition.
The sensitivity is \code{NA} for every design point.
}
\description{
Drops insignificant design points and merges design points in a certain neighbourhood.
}
\examples{
#TODO
}
\seealso{
\code{\link{FedorovWynn}}
}

