% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{roworder}
\alias{roworder}
\title{Matrix Ordering Permutation}
\usage{
roworder(x, ...)
}
\arguments{
\item{x}{a matrix.}

\item{...}{other arguments passed to \code{order}.}
}
\value{
\code{roworder} returns an integer vector.
}
\description{
\code{roworder} returns a permutation which rearranges the rows of its first argument into ascending order.
}
\examples{
x = expand.grid(1:3, 1:2, 3:1)
x = x[sample(seq1(1, nrow(x)), nrow(x)),]
x

ord = roworder(x)
ord

x[ord,]
}
\seealso{
\code{\link[base]{order}}
}

