% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{rowsduplicated}
\alias{rowsduplicated}
\title{Determine Duplicate Rows}
\usage{
rowsduplicated(x)
}
\arguments{
\item{x}{a row matrix of doubles.}
}
\value{
\code{rowsduplicated} returns a logical vector with one element for each row.
}
\description{
\code{rowsduplicated} determines which rows of a matrix are duplicates of rows with smaller subscripts, and returns a logical vector indicating which rows are duplicates.
}
\details{
\code{rowsduplicated} uses compiled C-code.
}
\seealso{
\code{\link[base]{duplicated}}
}
