% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{seq1}
\alias{seq1}
\title{Sequence Generation}
\usage{
seq1(from, to, by = 1)
}
\arguments{
\item{from,to,by}{see \code{\link[base]{seq}} in \pkg{base}.}
}
\value{
\code{seq1} returns either \code{integer(0)} if range is empty or what an appropriate call to \code{base::seq} returns otherwise.

See examples below.
}
\description{
Similar to \code{base::seq}, however \code{by} is strictly 1 by default and \code{integer(0)} is returned if range is empty.
}
\examples{
library(docopulae)

## by = 1
seq1(1, 3)
seq1(3, 1)
3:1

## by = -3
seq1(5, -2, -3)
seq1(5, 5, -3)
seq1(5, 3, -3)
}
\seealso{
\code{\link[base]{seq}} in \pkg{base}
}

