% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base.R
\name{seqi}
\alias{seqi}
\title{Sequence Generation}
\usage{
seqi(from, to, by = 1)
}
\arguments{
\item{from,to,by}{see \code{\link[base]{seq}} in \pkg{base}.}
}
\value{
\code{seqi} returns either \code{integer(0)} if no closed interval is contained in [\code{from}, \code{to}] (may be empty) or what an appropriate call to \code{base::seq} returned otherwise.

See examples below.
}
\description{
Similar to \code{base::seq}, however \code{by} is strictly 1 by default (not changing with \code{from}, \code{to}) and every closed interval [result, result + \code{by}] is contained in [\code{from}, \code{to}].
}
\examples{
library(docopulae)

## by = 1
seqi(1, 3)
seqi(3, 1)
3:1

## by = -3
seqi(5, -2, -3) # [5, 3], [2, 0]
seqi(5, 4, -3)
seqi(5, 3, -3)
}
\seealso{
\code{\link[base]{seq}} in \pkg{base}
}

