% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update.design}
\alias{update.design}
\title{Update Design}
\usage{
\method{update}{design}(des)
}
\arguments{
\item{des}{some design.}
}
\value{
\code{update.design} returns an object of \code{class} \code{"design"}. See \code{FedorovWynn} for its structural definition.
}
\description{
\code{update.design} updates the underlying model of some design to ensure the existence of Fisher information matrices for each point in the design.
It also updates the corresponding sensitivities.
}
\examples{
#TODO
}
\seealso{
\code{\link{reduce}}, \code{\link{update.parm}}, \code{\link{FedorovWynn}}
}

