% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update.param}
\alias{update.param}
\title{Update Parametric Model}
\usage{
\method{update}{param}(mod, x)
}
\arguments{
\item{mod}{an object of class \code{"param"}.}

\item{x}{a row matrix of points to add.
The number of columns shall be equal to \code{mod$dDim}.}
}
\value{
\code{update.param} returns an object of \code{class} \code{"param"}.
}
\description{
\code{update.param} evaluates the Fisher information (if necessary) and returns an updated model object.
}
\seealso{
\code{\link{param}}
}

