% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update.parm}
\alias{update.parm}
\title{Update Parametric Model}
\usage{
\method{update}{parm}(mod, x)
}
\arguments{
\item{mod}{an object of class \code{"parm"}.}

\item{x}{a row matrix of points to evaluate the Fisher information at.
The number of columns shall be equal to \code{mod$dDim}.}
}
\value{
\code{update.parm} returns an object of \code{class} \code{"parm"}.
}
\description{
\code{update.parm} evaluates the Fisher information if necessary and returns the updated model object.
}
\seealso{
\code{\link{parm}}
}

