% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update_reference}
\alias{update_reference}
\title{Update Reference Design}
\usage{
update_reference(ref, other)
}
\arguments{
\item{ref}{some design.}

\item{other}{either a single design or a list structure of designs.}
}
\value{
\code{update_reference} returns an object of \code{class} \code{"design"}.
See \code{FedorovWynn} for its structural definition.
}
\description{
\code{update_reference} updates the underlying model of some design to ensure the existence of Fisher information matrices for each point in each other design.
}
\examples{
#TODO
}
\seealso{
\code{\link{Defficiency}}, \code{\link{update.design}}, \code{\link{update.parm}}
}

