% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docopulae.R
\docType{package}
\name{docopulae}
\alias{docopulae}
\alias{docopulae-package}
\title{Design of Experiments with Copulas}
\description{
docopulae provides a set of functions to define parametric models by their joint densities (likelihood), evaluate the Fisher information, find D- and Ds-optimal designs and eventually compute common statistics. A crucial aspect is modularity. At any point the user may employ specialized solutions. The package also consists of a number of convenience functions, some general purpose, some specific to the task.
}
\details{
TODO
}

