% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{expr2f}
\alias{expr2f}
\title{Expression To Function}
\usage{
expr2f(x, map = NULL, yMap = NULL, thetaMap = NULL)
}
\arguments{
\item{x}{an expression.}

\item{map}{a named list of character strings defining left assignments (\code{a="b"} => \code{a <- b}).}

\item{yMap}{like \code{map} with \code{a=b} resolving to \code{a <- y[b]}.}

\item{thetaMap}{like \code{map} with \code{a=b} resolving to \code{a <- theta[[b]]}.}
}
\value{
\code{expr2f} returns \code{function(y, theta, ...)}, where \code{theta} is a list of parameters.
It evaluates expression \code{x}.
}
\description{
\code{expr2f} turns an expression into \code{function(y, theta, ...)}.
}
\examples{
## see examples for param
}
\seealso{
\code{\link{buildf}}, \code{\link{fisherI}}
}

