% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{fisherI}
\alias{fisherI}
\title{Fisher Information}
\usage{
fisherI(f, theta, names, dlogf = NULL, d2logf = NULL, yspace = NULL, ...,
  yispaces = NULL, transformInf = F)
}
\arguments{
\item{f}{\code{function(y, theta, ...)}, a joint PDF.}

\item{theta}{a list of parameters.}

\item{names}{a vector of names or indices, the subset of parameters.}

\item{dlogf}{\code{function(y, theta, i, ...)} which evaluates to the first derivative of \code{log(f)} with respect to \code{theta[[i]]}.}

\item{d2logf}{\code{function(y, theta, i, j, ...)} which evaluates to the second derivative of \code{log(f)} with respect to \code{theta[[i]]} and \code{theta[[j]]}.}

\item{yspace}{the support of \code{y}.
See \code{\link{nint_space}}.}

\item{...}{other arguments passed to \code{d2logf}.}

\item{yispaces}{see \code{\link{nint_integrate}}.}

\item{transformInf}{see \code{\link{nint_integrate}}.}
}
\value{
\code{fisherI} returns a named matrix, the Fisher information.
}
\description{
\code{fisherI} utilizes \code{nint_integrate} to evaluate the Fisher information.
}
\details{
Either \code{dlogf} xor \code{d2logf} shall be given.
}
\examples{
#TODO
}
\seealso{
\code{\link{buildf}}, \code{\link{numDerivLogf}}, \code{\link{DerivLogf}}, \code{\link{nint_space}}, \code{\link{nint_integrate}}
}

