% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getM}
\alias{getM}
\title{Get Fisher Information}
\usage{
getM(mod, des)
}
\arguments{
\item{mod}{a model.}

\item{des}{a design.}
}
\value{
\code{getM} returns a named matrix, the Fisher information.
}
\description{
\code{getM} returns the Fisher information corresponding to a model and a design.
}
\examples{
## see examples for param

}
\seealso{
\code{\link{param}}, \code{\link{design}}
}

