% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{getM}
\alias{getM}
\title{Get Fisher Information}
\usage{
getM(mod, des)
}
\arguments{
\item{mod}{some model.}

\item{des}{some design.}
}
\value{
\code{getM} returns a named matrix, the Fisher information.
}
\description{
\code{getM} returns the Fisher information corresponding to some model and some design.
}
\examples{
## see examples for param
}
\seealso{
\code{\link{param}}, \code{\link{design}}
}

