% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_gridDim}
\alias{nint_gridDim}
\title{Grid Dimension}
\usage{
nint_gridDim(x)
}
\arguments{
\item{x}{a vector of any type.}
}
\value{
\code{nint_scatDim} returns its argument with the dimension type attribute set to \code{nint_TYPE_GRID_DIM}.
}
\description{
\code{nint_gridDim} is defined by a sequence of values.
Together with other grid dimensions it defines a dense grid.
}
\details{
Imagine using \code{expand.grid} to create a row matrix of points.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}
