% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_intvDim}
\alias{nint_intvDim}
\title{Interval Dimension}
\usage{
nint_intvDim(x, b = NULL)
}
\arguments{
\item{x}{either a single numeric, the lower bound, or a vector of length 2, the lower and upper bound.}

\item{b}{the upper bound if \code{x} is the lower bound.}
}
\value{
\code{nint_intvDim} returns a vector of length 2 with the dimension type attribute set to \code{nint_TYPE_INTV_DIM}.
}
\description{
\code{nint_intvDim} defines a fixed interval.
The bounds may be (negative) \code{Inf}.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}
