% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nint.R
\name{nint_scatDim}
\alias{nint_scatDim}
\title{Scatter Dimension}
\usage{
nint_scatDim(x)
}
\arguments{
\item{x}{a vector of any type.}
}
\value{
\code{nint_scatDim} returns its argument with the dimension type attribute set to \code{nint_TYPE_SCAT_DIM}.
}
\description{
\code{nint_scatDim} is defined by a sequence of values.
Together with other scatter dimensions it defines a sparse grid.
}
\details{
Imagine using \code{cbind} to create a row matrix of points.
}
\seealso{
\code{\link{nint_TYPE}}, \code{\link{nint_space}}
}
