% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{reduce}
\alias{reduce}
\title{Reduce Design}
\usage{
reduce(des, distMax, wMin = 1e-06)
}
\arguments{
\item{des}{some design.}

\item{distMax}{maximum euclidean distance between points to be merged.}

\item{wMin}{minimum weight a design point shall have to be considered significant.}
}
\value{
\code{reduce} returns an object of \code{class} \code{"desigh"}.
See \code{FedorovWynn} for its structural definition.
The sensitivity is set to \code{NA} for all design points.
}
\description{
\code{reduce} drops insignificant design points and merges design points in a certain neighbourhood.
}
\examples{
## see examples for param
}
\seealso{
\code{\link{FedorovWynn}}, \code{\link{update.desigh}}, \code{\link{plot.desigh}}
}

