% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{seq1}
\alias{seq1}
\title{Sequence Generation}
\usage{
seq1(from, to, by = 1)
}
\arguments{
\item{from, to, by}{see \code{\link[base]{seq}}.}
}
\value{
\code{seq1} returns either \code{integer(0)} if range is empty or what an appropriate call to \code{seq} returns otherwise.

See examples below.
}
\description{
\code{seq1} is similar to \code{seq}, however \code{by} is strictly \code{1} by default and \code{integer(0)} is returned if the range is empty.
}
\examples{
seq1(1, 3)
seq1(3, 1) # different from seq
seq(3, 1)
3:1

seq1(5, 1, -3)
}
\seealso{
\code{\link[base]{seq}}
}

