% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update.design}
\alias{update.design}
\title{Update Design}
\usage{
\method{update}{design}(des)
}
\arguments{
\item{des}{some design.}
}
\value{
\code{update.design} returns an object of \code{class} \code{"design"}.
See \code{FedorovWynn} for its structural definition.
}
\description{
\code{update.design} updates the underlying model and the sensitivities.
Usually this is necessary for custom or transformed (e.g. reduced) designs.
}
\examples{
## see examples for param
}
\seealso{
\code{\link{reduce}}, \code{\link{update.param}}, \code{\link{Defficiency}}, \code{\link{FedorovWynn}}, \code{\link{getM}}
}

