% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{update.param}
\alias{update.param}
\title{Update Parametric Model}
\usage{
\method{update}{param}(object, x, ...)
}
\arguments{
\item{object}{some model.}

\item{x}{a row matrix of points.
The number of columns shall be equal to \code{object$dDim}.}

\item{...}{ignored.}
}
\value{
\code{update.param} returns an object of \code{class} \code{"param"}.
}
\description{
\code{update.param} evaluates the Fisher information at uncharted points and returns an updated model object.
}
\examples{
## see examples for param
}
\seealso{
\code{\link{param}}, \code{\link{update.desigh}}, \code{\link{update_reference}}, \code{\link{getM}}
}

