% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update.param}
\alias{update.param}
\title{Update Parametric Model}
\usage{
\method{update}{param}(object, x, ...)
}
\arguments{
\item{object}{a model.}

\item{x}{either a row matrix of points or a design, or a list structure of matrices or designs.
The number of columns/the dimensionality of the design space shall be equal to \code{ncol(object$x)}.}

\item{...}{ignored.}
}
\value{
\code{update.param} returns an object of \code{class} \code{"param"}.
See \code{\link{param}} for its structural definition.
}
\description{
\code{update.param} evaluates the Fisher information at uncharted points and returns an updated model object.
}
\details{
When the user interrupts execution, the function returns a partially updated model object.
}
\examples{
## see examples for param

}
\seealso{
\code{\link{param}}, \code{\link{grow.grid}}, \code{\link{design}}
}
